requirements.txt
streamlit==1.38.0
opencv-python==4.8.0.76
ultralytics==8.3.10
torch==2.2.2
pytesseract==0.3.10
numpy==1.24.0
pillow==10.0.0